fixed laggy scroll in mobile (last attempt, if fail, I give up)
The issue is, I am using dvh to take full screen size which is the best option for mobile responsive site. But I am also using the same div as relative to Hero Orbits. Now, whenever browser's address bar or bottom bar hides/shows the dvh re-calculate the viewport which makes the Orbits re-paint in the screen. Thus user see a flickering. I've some idea how it can be fixed, but I am not skilled enough in css to actually implement them. So bye bye :)